// apps/gq/src/app/shared/services/rest/user-settings/models/user-settings-keys.enum.ts
export enum UserSettingsKeys {
  LOCALE = 'locale',
  LANGUAGE = 'language',
  GQ_CASE_OVERVIEW_STATE = 'GQ_CASE_OVERVIEW_STATE',
  GQ_PROCESS_CASE_STATE = 'GQ_PROCESS_CASE_STATE',
  GQ_REFERENCE_PRICING_TABLE_STATE = 'GQ_REFERENCE-PRICING-TABLE_STATE',
  GQ_SAP_PRICE_DETAILS_STATE = 'GQ_SAP-PRICE-DETAILS_STATE',
  GQ_TRANSACTIONS_STATE = 'GQ_TRANSACTIONS_STATE',
  GQ_SEARCH_MATERIAL_RESULTS_TABLE_STATE = 'GQ_SEARCH-MATERIAL-RESULTS-TABLE_STATE',
  GQ_SEARCH_CASES_RESULTS_TABLE_STATE = 'GQ_SEARCH-CASES-RESULTS-TABLE_STATE',
}